Fixed previous commit message: Performance
The previous commit message was for PLANVIZ, these are the changes
for plan-schema:

Added a slot for nodes and edges called "number" which is a vector
of integers describing the position of each element in the graph.
Using this number it is much more performant to determine if one
element is within the scope of a given node.

Used the new element numbering to improve the performance
of merging HTN and TPN plans.

Updated dependencies

Signed-off-by: Tom Marble <tmarble@info9.net>